requirements.txt
Pillow
numpy
opencv-python
PyQt5
scikit-image
scipy
pyopencl